chat-api/requirements.txt
python-dotenv 
google-generativeai 
langchain 
langchain-chroma 
langchain-google-genai 
langchain-huggingface 
pypdf 
sentence-transformers
pdf2image 
pytesseract 
unstructured 
pdfplumber
PyMuPDF==1.23.8  # for fitz
opencv-python==4.8.1.78
# Data Processing
pandas==2.1.4
numpy==1.26.2
pyyaml==6.0.1
chromadb==0.4.22
openai
faiss-cpu
fastapi
uvicorn
langchain_community
# code formatting
black
